# [NoQA] Improve performance of isConsecutiveActionMadeByPreviousActor